fix: Build config in main
